Stable version for stage construction1
Usage:
/ASP/admin for managing invitation, token is redirect to the console, use the token to go to registration page
/login/ for login, you should be redirected to the corresponding page after successful login

Please note that the queries for creating clinic manager, clinic, etc in the data.py should not be used anymore. I don't have time to change the lagacy code in the Dispatch part, so there's a little trick to do in order for the code in the Dispatch part to work. But that shouldn't be the problem as dispatching isn't involved in this stage.